requirements.txt
Django==4.2.7
mysqlclient==2.2.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
qiniu==7.12.1
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
gunicorn==21.2.0